:recycle: Refactor all code and remove patch classifier. Replace it with patch predictor.
