+++ tiatoolbox/models/engine/patch_predictor.py
@@ -2,22 +2,91 @@
 
 from __future__ import annotations
 
+import math
 from typing import TYPE_CHECKING
 
-from .patch_classifier import PatchClassifier
+import zarr
+from typing_extensions import Unpack
+
+from .engine_abc import EngineABC, EngineABCRunParams
 
 if TYPE_CHECKING:  # pragma: no cover
+    import os
     from pathlib import Path
 
+    import numpy as np
+
+    from tiatoolbox.annotation import AnnotationStore
+    from tiatoolbox.models.engine.io_config import ModelIOConfigABC
     from tiatoolbox.models.models_abc import ModelABC
+    from tiatoolbox.wsicore import WSIReader
 
 
-class PatchPredictor(PatchClassifier):
-    r"""Patch level predictor for digital histology images.
+class PredictorRunParams(EngineABCRunParams):
+    """Class describing the input parameters for the :func:`EngineABC.run()` method.
 
-    .. warning::
-        PatchPredictor will be deprecated in v2.1 of TIAToolbox. Please use
-        `PatchClassifier` instead.
+    Attributes:
+        batch_size (int):
+            Number of image patches to feed to the model in a forward pass.
+        cache_mode (bool):
+            Whether to run the Engine in cache_mode. For large datasets,
+            we recommend to set this to True to avoid out of memory errors.
+            For smaller datasets, the cache_mode is set to False as
+            the results can be saved in memory.
+        cache_size (int):
+            Specifies how many image patches to process in a batch when
+            cache_mode is set to True. If cache_size is less than the batch_size
+            batch_size is set to cache_size.
+        class_dict (dict):
+            Optional dictionary mapping classification outputs to class names.
+        device (str):
+            Select the device to run the model. Please see
+            https://pytorch.org/docs/stable/tensor_attributes.html#torch.device
+            for more details on input parameters for device.
+        ioconfig (ModelIOConfigABC):
+            Input IO configuration (:class:`ModelIOConfigABC`) to run the Engine.
+        return_labels (bool):
+            Whether to return the labels with the predictions.
+        num_loader_workers (int):
+            Number of workers used in :class:`torch.utils.data.DataLoader`.
+        num_post_proc_workers (int):
+            Number of workers to postprocess the results of the model.
+        output_file (str):
+            Output file name to save "zarr" or "db". If None, path to output is
+            returned by the engine.
+        patch_input_shape (tuple):
+            Shape of patches input to the model as tuple of height and width (HW).
+            Patches are requested at read resolution, not with respect to level 0,
+            and must be positive.
+        resolution (Resolution):
+            Resolution used for reading the image. Please see
+            :class:`WSIReader` for details.
+        return_probabilities (bool):
+                Whether to return per-class probabilities.
+        scale_factor (tuple[float, float]):
+            The scale factor to use when loading the
+            annotations. All coordinates will be multiplied by this factor to allow
+            conversion of annotations saved at non-baseline resolution to baseline.
+            Should be model_mpp/slide_mpp.
+        stride_shape (tuple):
+            Stride used during WSI processing. Stride is
+            at requested read resolution, not with respect to
+            level 0, and must be positive. If not provided,
+            `stride_shape=patch_input_shape`.
+        units (Units):
+            Units of resolution used for reading the image. Choose
+            from either `level`, `power` or `mpp`. Please see
+            :class:`WSIReader` for details.
+        verbose (bool):
+            Whether to output logging information.
+
+    """
+
+    return_probabilities: bool
+
+
+class PatchPredictor(EngineABC):
+    r"""Patch level prediction for digital histology images.
 
     The models provided by TIAToolbox should give the following results:
 
@@ -277,15 +346,6 @@ class PatchPredictor(PatchClassifier):
         verbose: bool = True,
     ) -> None:
         """Initialize :class:`PatchPredictor`."""
-        from tiatoolbox import logger
-
-        msg = (
-            "PatchPredictor will be deprecated in v2.1 of TIAToolbox."
-            "Please use PatchClassifier instead."
-        )
-        logger.warning(
-            msg=msg,
-        )
         super().__init__(
             model=model,
             batch_size=batch_size,
@@ -294,4 +354,197 @@ class PatchPredictor(PatchClassifier):
             weights=weights,
             device=device,
             verbose=verbose,
+        )
+
+    def post_process_cache_mode(
+        self: PatchPredictor,
+        raw_predictions: Path,
+        **kwargs: Unpack[PredictorRunParams],
+    ) -> Path:
+        """Returns an array from raw predictions."""
+        return_probabilities = kwargs.get("return_probabilities")
+        zarr_group = zarr.open(str(raw_predictions), mode="r+")
+
+        num_iter = math.ceil(len(zarr_group["probabilities"]) / self.batch_size)
+        start = 0
+        for _ in range(num_iter):
+            # Probabilities for post-processing
+            probabilities = zarr_group["probabilities"][start : start + self.batch_size]
+            start = start + self.batch_size
+            predictions = self.model.postproc_func(
+                probabilities,
+            )
+            if "predictions" in zarr_group:
+                zarr_group["predictions"].append(predictions)
+                continue
+
+            zarr_dataset = zarr_group.create_dataset(
+                name="predictions",
+                shape=predictions.shape,
+                compressor=zarr_group["probabilities"].compressor,
+            )
+            zarr_dataset[:] = predictions
+
+        if return_probabilities is not False:
+            return raw_predictions
+
+        del zarr_group["probabilities"]
+
+        return raw_predictions
+
+    def post_process_patches(
+        self: PatchPredictor,
+        raw_predictions: dict | Path,
+        **kwargs: Unpack[PredictorRunParams],
+    ) -> dict | Path:
+        """Post-process raw patch predictions from inference.
+
+        The output of :func:`infer_patches()` with patch prediction information will be
+        post-processed using this function. The processed output will be saved in the
+        respective input format. If `cache_mode` is True, the function processes the
+        input using zarr group with size specified by `cache_size`.
+
+        Args:
+            raw_predictions (dict | Path):
+                A dictionary or path to zarr with patch prediction information.
+            **kwargs (PredictorRunParams):
+                Keyword Args to update setup_patch_dataset() method attributes. See
+                :class:`PredictorRunParams` for accepted keyword arguments.
+
+        Returns:
+            dict or Path:
+                Returns patch based output after post-processing. Returns path to
+                saved zarr file if `cache_mode` is True.
+
+        """
+        return_probabilities = kwargs.get("return_probabilities")
+        if self.cache_mode:
+            return self.post_process_cache_mode(raw_predictions, **kwargs)
+
+        probabilities = raw_predictions.get("probabilities")
+
+        predictions = self.model.postproc_func(
+            probabilities,
+        )
+
+        raw_predictions["predictions"] = predictions
+
+        if return_probabilities is not False:
+            return raw_predictions
+
+        del raw_predictions["probabilities"]
+
+        return raw_predictions
+
+    def post_process_wsi(
+        self: PatchPredictor,
+        raw_predictions: dict | Path,
+        **kwargs: Unpack[PredictorRunParams],
+    ) -> dict | Path:
+        """Post process WSI output.
+
+        Takes the raw output from patch predictions and post-processes it to improve the
+        results e.g., using information from neighbouring patches.
+
+        """
+        return self.post_process_cache_mode(raw_predictions, **kwargs)
+
+    def run(
+        self: PatchPredictor,
+        images: list[os | Path | WSIReader] | np.ndarray,
+        masks: list[os | Path] | np.ndarray | None = None,
+        labels: list | None = None,
+        ioconfig: ModelIOConfigABC | None = None,
+        *,
+        patch_mode: bool = True,
+        save_dir: os | Path | None = None,  # None will not save output
+        overwrite: bool = False,
+        output_type: str = "dict",
+        **kwargs: Unpack[PredictorRunParams],
+    ) -> AnnotationStore | Path | str | dict:
+        """Run the engine on input images.
+
+        Args:
+            images (list, ndarray):
+                List of inputs to process. when using `patch` mode, the
+                input must be either a list of images, a list of image
+                file paths or a numpy array of an image list.
+            masks (list | None):
+                List of masks. Only utilised when patch_mode is False.
+                Patches are only generated within a masked area.
+                If not provided, then a tissue mask will be automatically
+                generated for whole slide images.
+            labels (list | None):
+                List of labels. Only a single label per image is supported.
+            patch_mode (bool):
+                Whether to treat input image as a patch or WSI.
+                default = True.
+            ioconfig (IOPatchPredictorConfig):
+                IO configuration.
+            save_dir (str or pathlib.Path):
+                Output directory to save the results.
+                If save_dir is not provided when patch_mode is False,
+                then for a single image the output is created in the current directory.
+                If there are multiple WSIs as input then the user must provide
+                path to save directory otherwise an OSError will be raised.
+            overwrite (bool):
+                Whether to overwrite the results. Default = False.
+            output_type (str):
+                The format of the output type. "output_type" can be
+                "zarr" or "AnnotationStore". Default value is "zarr".
+                When saving in the zarr format the output is saved using the
+                `python zarr library <https://zarr.readthedocs.io/en/stable/>`__
+                as a zarr group. If the required output type is an "AnnotationStore"
+                then the output will be intermediately saved as zarr but converted
+                to :class:`AnnotationStore` and saved as a `.db` file
+                at the end of the loop.
+            **kwargs (PredictorRunParams):
+                Keyword Args to update :class:`EngineABC` attributes during runtime.
+
+        Returns:
+            (:class:`numpy.ndarray`, dict):
+                Model predictions of the input dataset. If multiple
+                whole slide images are provided as input,
+                or save_output is True, then results are saved to
+                `save_dir` and a dictionary indicating save location for
+                each input is returned.
+
+                The dict has the following format:
+
+                - img_path: path of the input image.
+                - raw: path to save location for raw prediction,
+                  saved in .json.
+
+        Examples:
+            >>> wsis = ['wsi1.svs', 'wsi2.svs']
+            >>> class PatchPredictor(EngineABC):
+            >>> # Define all Abstract methods.
+            >>>     ...
+            >>> predictor = PatchPredictor(model="resnet18-kather100k")
+            >>> output = predictor.run(image_patches, patch_mode=True)
+            >>> output
+            ... "/path/to/Output.db"
+            >>> output = predictor.run(
+            >>>     image_patches,
+            >>>     patch_mode=True,
+            >>>     output_type="zarr")
+            >>> output
+            ... "/path/to/Output.zarr"
+            >>> output = predictor.run(wsis, patch_mode=False)
+            >>> output.keys()
+            ... ['wsi1.svs', 'wsi2.svs']
+            >>> output['wsi1.svs']
+            ... {'/path/to/wsi1.db'}
+
+        """
+        return super().run(
+            images=images,
+            masks=masks,
+            labels=labels,
+            ioconfig=ioconfig,
+            patch_mode=patch_mode,
+            save_dir=save_dir,
+            overwrite=overwrite,
+            output_type=output_type,
+            **kwargs,
         )
